extract health controller from game controller
