+++ src/main/java/com/example/tankmark1/GameController.java
@@ -52,7 +52,7 @@ public class GameController extends Pane {
     private boolean isRobot=false;
     private ComputerTank computerTank;
     private List<Tank> allTanks=new ArrayList<>();
-    private Text pauseText;
+    HealthController healthController;
 
     public GameController(int numPlayers, String selectedWeapon, String selectedMap, boolean soundOn, TankGame mainApp,String level) {
         this.numPlayers = numPlayers;
@@ -63,25 +63,15 @@ public class GameController extends Pane {
         this.level=level;
         soundManager = new GameSoundManager();
 
+
         setUpMap();
         setUpTanks();
+        healthController=new HealthController(this,tank1,tank2);
 
-        setUpHealthBars();
+        healthController.setUpHealthBars();
         startCountdown();
-        setUpPauseText();
-
-
     }
 
-    private void setUpPauseText() {
-        pauseText = new Text("Game Paused\nPress 'R' to Resume or 'M' to return to Menu");
-        pauseText.setFont(new Font(50));
-        pauseText.setFill(Color.RED);
-        pauseText.setLayoutX(400); // Center horizontally
-        pauseText.setLayoutY(400); // Center vertically
-        pauseText.setVisible(false);  // Initially hidden
-        this.getChildren().add(pauseText);
-    }
     private void startCountdown() {
 
         countdownText = new Text();
@@ -135,101 +125,7 @@ public class GameController extends Pane {
         countdown.setCycleCount(1);
         countdown.play();
     }
-    private void setUpHealthBars() {
-        // Player 1 Label
-        Text player1Label = new Text("Player 1");
-        player1Label.setFont(new Font(18));
-        player1Label.setFill(Color.RED);
-        player1Label.setLayoutX(10);  // Position above Player 1's health bar
-        player1Label.setLayoutY(25);
 
-        // Player 1 Health Bar
-        healthBar1 = new ProgressBar(1);
-        healthBar1.setStyle("-fx-accent: red;");
-        healthText1 = new Text("100%");
-        HBox player1HealthBox = new HBox(5, healthBar1, healthText1);
-        player1HealthBox.setLayoutX(10);
-        player1HealthBox.setLayoutY(40);
-
-        // Player 2 Label
-        Text player2Label = new Text("Player 2");
-        player2Label.setFont(new Font(18));
-        player2Label.setFill(Color.BLUE);
-        player2Label.setLayoutX(1400);  // Position above Player 2's health bar
-        player2Label.setLayoutY(25);
-
-        // Player 2 Health Bar
-        healthBar2 = new ProgressBar(1);
-        healthBar2.setStyle("-fx-accent: BLUE;");
-        healthText2 = new Text("100%");
-        HBox player2HealthBox = new HBox(5, healthText2, healthBar2);
-        player2HealthBox.setLayoutX(1400);
-        player2HealthBox.setLayoutY(40);
-
-        // Add to the scene
-        this.getChildren().addAll(player1Label, player1HealthBox, player2Label, player2HealthBox);
-
-        // Winner Text
-        winnerText = new Text();
-        winnerText.setFill(Color.RED);
-        winnerText.setLayoutX(350);
-        winnerText.setLayoutY(50);
-        winnerText.setVisible(false);
-        this.getChildren().add(winnerText);
-    }
-
-    private void setsb() {
-        sb1 = new ProgressBar(1);
-        sb1.setStyle("-fx-accent: black;");
-        st1 = new Text("100%");
-
-        // Position Player 1’s health on the left
-        HBox player1sb = new HBox(5, sb1, st1);
-        player1sb.setLayoutX(10);
-        player1sb.setLayoutY(10);
-
-        sb2 = new ProgressBar(1);
-        sb2.setStyle("-fx-accent: black;");
-        st2 = new Text("100%");
-
-        // Position Player 2’s health on the right
-        HBox player2sb = new HBox(5, st2, sb2);
-        player2sb.setLayoutX(1400); // Adjust based on scene width
-        player2sb.setLayoutY(10);
-        this.getChildren().addAll(player1sb,player2sb);
-
-        winnerText = new Text();
-        winnerText.setFill(Color.RED);
-        winnerText.setLayoutX(350);
-        winnerText.setLayoutY(50);
-        winnerText.setVisible(false);
-        this.getChildren().add(winnerText);
-    }
-
-//    public void updatesb() {
-//        // Update Player 1 health bar and text
-//        double s1 = tank1.getHealth();
-//        sb1.setProgress(s1 / 100.0);
-//        st1.setText((int) s1 + "%");
-//
-//        // Update Player 2 health bar and text
-//        double s2 = tank2.getHealth();
-//        sb2.setProgress(s2 / 100.0);
-//       st2.setText((int) s2 + "%");
-//    }
-
-
-    public void updateHealthBars() {
-        // Update Player 1 health bar and text
-        double health1 = tank1.getHealth();
-        healthBar1.setProgress(health1 / 100.0);
-        healthText1.setText((int) health1 + "%");
-
-        // Update Player 2 health bar and text
-        double health2 = tank2.getHealth();
-        healthBar2.setProgress(health2 / 100.0);
-        healthText2.setText((int) health2 + "%");
-    }
 
     public void checkForWin() {
         if (tank1.isDestroyed()) {
@@ -463,7 +359,7 @@ public class GameController extends Pane {
                     tank1.takeDamageToS(projectile.getDamage());
                 }
 //                updatesb();
-                updateHealthBars();
+                healthController.updateHealthBars();
                 removeProjectile(projectile);
                 checkForWin();
 
@@ -482,7 +378,7 @@ public class GameController extends Pane {
                     tank2.takeDamageToS(projectile.getDamage());
                 }
 //                updatesb();
-                updateHealthBars();
+                healthController.updateHealthBars();
                 removeProjectile(projectile);
                 checkForWin();
 
@@ -501,8 +397,8 @@ public class GameController extends Pane {
         double collisionX = (projectile.getX() + projectile.getX()) / 2;
         double collisionY = (projectile.getY() + projectile.getY()) / 2;
 
-        double adjustedX = collisionX - (projectile.getExplosionFrameWidth() / 2);  // Center the explosion
-        double adjustedY =collisionY - (projectile.getExplosionFrameHeight() / 2); // Center the explosion
+        double adjustedX = collisionX - (128/ 2);  // Center the explosion
+        double adjustedY =collisionY - (128 / 2); // Center the explosion
 //        double adjustedX = projectile.getX() - 128/2;
 //        double adjustedY = projectile.getY() - 128/2;
         Explosion explosion = new Explosion(
